Remove `base` (#127)
* Remove `base`

* flake8 + black
