--- batchglm/api/typing.py
@@ -1,2 +1,0 @@
-from batchglm.models.base.estimator import EstimatorBaseTyping
-from batchglm.models.base.input import InputDataBase
