--- batchglm/models/base/external.py
@@ -1,2 +1,0 @@
-import batchglm.data as data_utils
-import batchglm.pkg_constants as pkg_constants
