+++ batchglm/models/base_glm/external.py
@@ -1,3 +1,2 @@
 import batchglm.data as data_utils
-from batchglm.models.base import InputDataBase, _EstimatorBase, _ModelBase, _SimulatorBase
 from batchglm.utils.linalg import groupwise_solve_lm
